Tmp path for config test
- Otherwise, profiler will open empty files in the examples.
